requirements.txt
Django>=4.0.1,<4.1
djangorestframework==3.14.0
mysqlclient==2.1.1
mysql-connector
daphne
channels
# Redis support
redis
django-redis==5.2.0
channels_redis
psycopg2
